Add basic support for populating database from tsv files
Modified build_database.py script to accept db_config flag allowing specification
of a json config file pointing at the location of two tsv files, one with nodes
information and one with edges information. Running the script with this flag
causes the tsv files to be ingested and loaded into a sqlite database.

Added rudimentary code for processing a Translator Reasoner Standard API query and
returning a list of edges for which either the subject or object node matches the
first node curie encountered in the node list. See main README for the repo to
see syntax for running the build_database.py script. Add --db_config as flag and
provide path to a json config file structured as follows:

{
	"dataset1": {
		"source_dir": "source-data-folder",
		"nodes_file": "source-data-nodes.tsv",
		"edges_file": "source-data-edges.tsv"
	}
}

The source tsv files are expected to be in the folder (relative path) specified by
"source_dir".